% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{call_cached}
\alias{call_cached}
\title{call_cached}
\usage{
call_cached(fun, cache_string = Sys.Date(), cache = memory_cache)
}
\arguments{
\item{fun}{The function to memoise}

\item{cache_string}{The string to use as the cache key}
}
\description{
Call a code block with memoisation
}
