% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{query_cdc_data}
\alias{query_cdc_data}
\title{query_cdc_data}
\usage{
query_cdc_data(dataset, query)
}
\arguments{
\item{dataset}{The dataset to query}

\item{query}{The query to pass to the dataset}
}
\description{
query_cdc_data
}
