% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{read_artifact}
\alias{read_artifact}
\title{read_artifact}
\usage{
read_artifact(
  name,
  extension = get_tabular_data_extension(),
  data_path = get_data_path()
)
}
\arguments{
\item{name}{The name of the artifact}
}
\description{
read_artifact
}
\examples{
read_artifact(iris)
read_artifact("iris")
}
