% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{write_artifact}
\alias{write_artifact}
\title{write_artifact}
\usage{
write_artifact(
  data,
  data_name = deparse(substitute(data)),
  data_path = get_data_path(),
  extension = get_tabular_data_extension()
)
}
\arguments{
\item{name}{The name of the artifact}
}
\description{
write_artifact
}
